this was utter garbage

+++ CG_CodeVsZombies2/Utils/AllowedDirections.cs
@@ -16,17 +16,17 @@ namespace CG_CodeVsZombies2.Utils
         {
             new Location(0, 0), // Stay
             new Location(1000, 0), // Right
-            new Location((int)Math.Round(1000 * Math.Cos(Math.PI / 4)),
-                (int)Math.Round(1000 * Math.Sin(Math.PI / 4))), // Upper right
+            new Location((short)Math.Round(1000 * Math.Cos(Math.PI / 4)),
+                (short)Math.Round(1000 * Math.Sin(Math.PI / 4))), // Upper right
             new Location(0, 1000), // Up
-            new Location((int)Math.Round(-1000 * Math.Cos(Math.PI / 4)),
-                (int)Math.Round(1000 * Math.Sin(Math.PI / 4))), // Upper left
+            new Location((short)Math.Round(-1000 * Math.Cos(Math.PI / 4)),
+                (short)Math.Round(1000 * Math.Sin(Math.PI / 4))), // Upper left
             new Location(-1000, 0), // Left
-            new Location((int)Math.Round(-1000 * Math.Cos(Math.PI / 4)),
-                (int)Math.Round(-1000 * Math.Sin(Math.PI / 4))), // Lower left
+            new Location((short)Math.Round(-1000 * Math.Cos(Math.PI / 4)),
+                (short)Math.Round(-1000 * Math.Sin(Math.PI / 4))), // Lower left
             new Location(0, -1000), // Down
-            new Location((int)Math.Round(1000 * Math.Cos(Math.PI / 4)),
-                (int)Math.Round(-1000 * Math.Sin(Math.PI / 4))) // Lower right
+            new Location((short)Math.Round(1000 * Math.Cos(Math.PI / 4)),
+                (short)Math.Round(-1000 * Math.Sin(Math.PI / 4))) // Lower right
         };
 
         public static Location GetRandom()
